requirements.txt
streamlit
boto3
python-dotenv
watchdog
pydantic
cryptography
asyncio
aiofiles 